Fix cache oblivious mult
