+++ src/strassen.h
@@ -4,81 +4,37 @@
 #include <cstddef>
 #include <utility>
 #include <vector>
+
 #include "matrix.h"
 #include "simple_multiplication.h"
 #include "view_matrix.h"
+#include "utils.h"
 
 namespace s_fast {
 
 namespace detail_strassen {
 
-using Index = typename Matrix<int>::Index;
-using Position = typename Matrix<int>::Position;
-
-inline Index GetNearestPowerOfTwo(Index number) {
-    Index power_of_two = 1;
-    while (power_of_two < number) {
-        power_of_two *= 2;
-    }
-    return power_of_two;
-}
-
-template <class ContainerType>
-struct BlockMatrix {
-    ContainerType left_top;
-    ContainerType right_top;
-    ContainerType left_bottom;
-    ContainerType right_bottom;
-};
-
-template <class InputContainer, class OuputContainer>
-BlockMatrix<OuputContainer> GetSubMatrixes(InputContainer& matrix) {
-    Index rows = GetNearestPowerOfTwo(matrix.Rows());
-    Index columns = GetNearestPowerOfTwo(matrix.Columns());
-
-    assert(rows > 1 && columns > 1);
-
-    return {.left_top = OuputContainer(matrix, {0, 0}, {rows / 2, columns / 2}),
-            .right_top = OuputContainer(matrix, {0, columns / 2}, {rows / 2, columns}),
-            .left_bottom = OuputContainer(matrix, {rows / 2, 0}, {rows, columns / 2}),
-            .right_bottom = OuputContainer(matrix, {rows / 2, columns / 2}, {rows, columns})};
-}
-
 template <class T>
-void SetSubMatrix(const Matrix<T>& from, Position begin, Matrix<T>* to) {
-    for (Index row = begin.row; row < std::min(to->Rows(), from.Rows() + begin.row); ++row) {
-        for (Index column = begin.column;
-             column < std::min(to->Columns(), from.Columns() + begin.column); ++column) {
-            (*to)(row, column) = from(row - begin.row, column - begin.column);
-        }
-    }
-}
-
-}  // namespace detail_strassen
-
-template <class T>
-Matrix<T> Strassen(const Matrix<T>& lhs, const Matrix<T>& rhs) {
-    using detail_strassen::BlockMatrix;
-    using detail_strassen::GetSubMatrixes;
-    using detail_strassen::SetSubMatrix;
+Matrix<T> Strassen(const ConstViewMatrix<T>& lhs, const ConstViewMatrix<T>& rhs) {
+    using utils::BlockMatrix;
+    using utils::GetSubMatrixes;
+    using utils::SetSubMatrix;
 
     assert(lhs.Columns() == rhs.Rows());
 
     if (std::min({lhs.Rows(), lhs.Columns(), rhs.Columns()}) <= 8) {
-        return SimpleMultiplication(lhs, rhs);
+        return SimpleMultiplication(GetMatrix(lhs), GetMatrix(rhs));
     }
 
-    auto lhs_sub = GetSubMatrixes<const Matrix<T>, ConstViewMatrix<T>>(lhs);
-    auto rhs_sub = GetSubMatrixes<const Matrix<T>, ConstViewMatrix<T>>(rhs);
+    auto lhs_sub = GetSubMatrixes<const ConstViewMatrix<T>, ConstViewMatrix<T>>(lhs);
+    auto rhs_sub = GetSubMatrixes<const ConstViewMatrix<T>, ConstViewMatrix<T>>(rhs);
 
     Matrix<T> m1 =
         Strassen(lhs_sub.left_top + lhs_sub.right_bottom, rhs_sub.left_top + rhs_sub.right_bottom);
-    Matrix<T> m2 =
-        Strassen(lhs_sub.left_bottom + lhs_sub.right_bottom, GetMatrix(rhs_sub.left_top));
-    Matrix<T> m3 = Strassen(GetMatrix(lhs_sub.left_top), rhs_sub.right_top - rhs_sub.right_bottom);
-    Matrix<T> m4 =
-        Strassen(GetMatrix(lhs_sub.right_bottom), rhs_sub.left_bottom - rhs_sub.left_top);
-    Matrix<T> m5 = Strassen(lhs_sub.left_top + lhs_sub.right_top, GetMatrix(rhs_sub.right_bottom));
+    Matrix<T> m2 = Strassen(lhs_sub.left_bottom + lhs_sub.right_bottom, rhs_sub.left_top);
+    Matrix<T> m3 = Strassen(lhs_sub.left_top, rhs_sub.right_top - rhs_sub.right_bottom);
+    Matrix<T> m4 = Strassen(lhs_sub.right_bottom, rhs_sub.left_bottom - rhs_sub.left_top);
+    Matrix<T> m5 = Strassen(lhs_sub.left_top + lhs_sub.right_top, rhs_sub.right_bottom);
     Matrix<T> m6 =
         Strassen(lhs_sub.left_bottom - lhs_sub.left_top, rhs_sub.left_top + rhs_sub.right_top);
     Matrix<T> m7 = Strassen(lhs_sub.right_top - lhs_sub.right_bottom,
@@ -101,6 +57,23 @@ Matrix<T> Strassen(const Matrix<T>& lhs, const Matrix<T>& rhs) {
     SetSubMatrix<T>(m1, {m7.Rows(), m7.Columns()}, &result);
 
     return result;
+}
+
+}  // namespace detail_strassen
+
+template <class T>
+Matrix<T> Strassen(const Matrix<T>& lhs, const Matrix<T>& rhs) {
+    return detail_strassen::Strassen(ConstViewMatrix<T>(lhs), ConstViewMatrix<T>(rhs));
+}
+
+template <class T>
+Matrix<T> Strassen(const ConstViewMatrix<T>& lhs, const Matrix<T>& rhs) {
+    return detail_strassen::Strassen(lhs, ConstViewMatrix<T>(rhs));
+}
+
+template <class T>
+Matrix<T> Strassen(const Matrix<T>& lhs, const ConstViewMatrix<T>& rhs) {
+    return detail_strassen::Strassen(ConstViewMatrix<T>(lhs), rhs);
 }
 
 }  // namespace s_fast
